refactor: move program-specific expiration logic to plugins and rename tray.py to main.py
