+++ plugins/hilton.py
@@ -13,6 +13,13 @@ class HiltonHonorsPlugin(ProviderPlugin):
     def plugin_id(self) -> str:
         return "hilton"
 
+    def calculate_expiration(self, balance: int, status: str, last_activity_date: datetime, has_exemption: bool = False) -> datetime:
+        from .base import add_months
+        return add_months(last_activity_date, 24)
+
+    def get_expiration_policy_description(self, status: str = None) -> str:
+        return "Points expire after 24 months of inactivity. Any earning or redemption transaction extends them."
+
     def _extract_data(self, sb) -> Tuple[Optional[int], Optional[str], Any]:
         """Extracts points balance, status, and last activity date from the Hilton Activity DOM."""
         balance, status, last_activity_date = None, None, None
