+++ plugins/marriott.py
@@ -12,6 +12,13 @@ class MarriottPlugin(ProviderPlugin):
     @property
     def plugin_id(self) -> str:
         return "marriott"
+
+    def calculate_expiration(self, balance: int, status: str, last_activity_date: datetime, has_exemption: bool = False) -> datetime:
+        from .base import add_months
+        return add_months(last_activity_date, 24)
+
+    def get_expiration_policy_description(self, status: str = None) -> str:
+        return "Points expire after 24 months of inactivity. Any earning or redemption transaction extends them."
         
     def _extract_from_datalayer(self, page_source: str) -> Tuple[Optional[int], Optional[str]]:
         """Extracts points balance and status directly from Marriott's global dataLayer."""
